Convert channel group dialogs to bottom sheets
