+++ app/src/main/java/com/github/libretube/ui/sheets/EditChannelGroupSheet.kt
@@ -1,34 +1,37 @@
-package com.github.libretube.ui.dialogs
+package com.github.libretube.ui.sheets
 
-import android.app.Dialog
 import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
 import androidx.core.view.isVisible
 import androidx.core.widget.addTextChangedListener
-import androidx.fragment.app.DialogFragment
 import androidx.fragment.app.activityViewModels
 import androidx.lifecycle.lifecycleScope
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.github.libretube.R
 import com.github.libretube.api.SubscriptionHelper
 import com.github.libretube.api.obj.Subscription
 import com.github.libretube.databinding.DialogEditChannelGroupBinding
 import com.github.libretube.db.obj.SubscriptionGroup
 import com.github.libretube.ui.adapters.SubscriptionGroupChannelsAdapter
 import com.github.libretube.ui.models.SubscriptionsViewModel
-import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 
-class EditChannelGroupDialog(
+class EditChannelGroupSheet(
     private var group: SubscriptionGroup,
     private val onGroupChanged: (SubscriptionGroup) -> Unit
-) : DialogFragment() {
+) : ExpandedBottomSheet() {
     private val subscriptionsModel: SubscriptionsViewModel by activityViewModels()
     private lateinit var binding: DialogEditChannelGroupBinding
     private var channels: List<Subscription> = listOf()
 
-    override fun onCreateDialog(savedInstanceState: Bundle?): Dialog {
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
         binding = DialogEditChannelGroupBinding.inflate(layoutInflater)
         binding.groupName.setText(group.name)
 
@@ -39,16 +42,18 @@ class EditChannelGroupDialog(
             showChannels(channels, it?.toString())
         }
 
-        return MaterialAlertDialogBuilder(requireContext())
-            .setTitle(R.string.edit_group)
-            .setNegativeButton(R.string.cancel, null)
-            .setPositiveButton(R.string.okay) { _, _ ->
-                group.name = binding.groupName.text.toString()
-                if (group.name.isBlank()) return@setPositiveButton
-                onGroupChanged(group)
-            }
-            .setView(binding.root)
-            .create()
+        binding.cancel.setOnClickListener {
+            dismiss()
+        }
+
+        binding.confirm.setOnClickListener {
+            group.name = binding.groupName.text.toString()
+            if (group.name.isBlank()) return@setOnClickListener
+            onGroupChanged(group)
+            dismiss()
+        }
+
+        return binding.root
     }
 
     private fun fetchSubscriptions() {
