+++ app/src/main/res/layout/subscription_group_row.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:orientation="horizontal"
-    android:paddingHorizontal="25dp"
+    android:paddingHorizontal="5dp"
     android:paddingVertical="10dp">
 
     <TextView
